The root file for led cube codes
I only started the naming and initialization of pins for the led cubes. I thought this can be a root file (where we can start with) of our led cube codes and every creative functions of our cube can be added here later. I do not have the cube with me so Alex can test it with this code and modify it if you find some of the names or variables misleading. I also use the variable ledPin to test out how the board works in general; I did not put other codes in, so when you put other codes, like sensors and joysticks or even the original code you guys found online for RGB led cube, please wrap them with comments so that I can easily comment them out or put them back into the codes for other testing purposes.